+++ @ornikar/eslint-config/tests/no-new-array.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const length = 10;
+// eslint-disable-next-line unicorn/no-new-array
+const array = new Array(length);
+
+exports.array = array;